requirements.txt
dj-database-url==0.5.0
Django==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
